requirements.txt
streamlit
ultralytics
pillow
Flask
flask-cors
gunicorn
opencv-python-headless
waitress